+++ sviper/lib/sviper.dart
@@ -1,2 +1,4 @@
 /// Annotation for module generator
+library;
+
 export 'src/annotation/sviper_annotation.dart';