+++ sviper/lib/sviper_module.dart
@@ -1,2 +1,4 @@
 /// Module parts
+library;
+
 export 'src/module/sviper.dart';