+++ sviper/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.1.10
+
+build: update dependencies
+docs: update the documentation
+
+## 1.1.9
+
+build: update dependencies
+
 ## 1.1.8
 
 docs: update the documentation